Holy fuck why.
This is the beginning of the rewrite for the checkers game.  Sourced and
everything.

Basically I see absolutely no way with the prior implementation that it's
even remotely possible to inject any websockets into it.  So making this
one and extending it to redraw the board on websocket calls, or that's the
idea anyway.  I believe in it more than the other one.